property_law_ai_clean/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
supabase==2.0.2
openai==1.3.5
reportlab==4.0.7
Pillow==10.1.0
aiofiles==23.2.1
httpx==0.24.1
pydantic==2.5.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
email-validator==2.2.0